requirements.txt
streamlit==1.37.1
requests==2.32.3
pandas==2.2.2
python-dotenv==1.0.1
plotly==5.23.0
numpy==2.0.1
schedule==1.2.0
sqlalchemy==2.0.23
